display world population

+++ src/main/java/com/napier/sem/App.java
@@ -77,7 +77,7 @@ public class App {
                     "select sum(Population) from(select population from world.city union all select population from world.country) as population";
             ResultSet rset = stmt.executeQuery(strSelect);
             if(rset.next()){
-                worldPopulation = rset.getInt("sum(Population)");
+                worldPopulation = rset.getLong("sum(Population)");
             }
         } catch (Exception e) {
             e.printStackTrace();
